moved logging to common
