+++ src/offspot_demo/__about__.py
@@ -1,1 +1,2 @@
+NAME = "demo"
 __version__ = "0.1.0-dev0"
